chore: move account fixtures
