+++ tests/conftest.py
@@ -2,6 +2,8 @@ import pytest
 from brownie import ZERO_ADDRESS, Contract
 from brownie_tokens import ERC20, MintableForkToken
 
+pytest_plugins = ["fixtures.accounts"]
+
 
 def pytest_addoption(parser):
     parser.addoption(
@@ -22,31 +24,6 @@ def pytest_addoption(parser):
 @pytest.fixture(autouse=True)
 def isolation_setup(fn_isolation):
     pass
-
-
-@pytest.fixture(scope="session")
-def alice(accounts):
-    yield accounts[0]
-
-
-@pytest.fixture(scope="session")
-def bob(accounts):
-    yield accounts[1]
-
-
-@pytest.fixture(scope="session")
-def charlie(accounts):
-    yield accounts[2]
-
-
-@pytest.fixture(scope="session")
-def dave(accounts):
-    yield accounts[3]
-
-
-@pytest.fixture(scope="session")
-def fee_receiver(accounts):
-    yield accounts[4]
 
 
 @pytest.fixture(scope="module")
